App: Routes: Add DNS Query Enpoint
The added endpoint takes a query parameter named "address" and uses
the same system utilities used by programs such as `ping` to resolve
the address. Will be used in conjunction with an error detection
script on the chromecast side to allow local intranet sites to be
loaded.

Should fully implement the server side of #18

Signed-off-by: Mark Stenglein <mark@stengle.in>