Build: JS: app.js: Add client side DNS checking
This patch allows the Chromecasts to pull from Local DNS if
the Channel's iFrames point to sites that are on the company intranet.

The chromecasts themselves are locked to google's DNS for security
reasons (keep you from serving hacked resources to them I suppose),
which is why intranet sites need to be pulled from the Multicast
server instead.

Signed-off-by: Mark Stenglein <mark@stengle.in>